+++ src/Handler/Exception/UnsupportedException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Gmponos\GuzzleLogger\Handler\Exception;
+
+/**
+ * @author George Mponos <gmponos@gmail.com>
+ */
+final class UnsupportedException extends \RuntimeException
+{
+}