+++ src/Middleware/LoggerMiddleware.php
@@ -3,43 +3,39 @@
 namespace Gmponos\GuzzleLogger\Middleware;
 
 use Closure;
+use Gmponos\GuzzleLogger\Handler\ArrayHandler;
+use Gmponos\GuzzleLogger\Handler\HandlerInterface;
 use GuzzleHttp\Exception\RequestException;
 use GuzzleHttp\TransferStats;
-use Psr\Http\Message\MessageInterface;
 use Psr\Http\Message\RequestInterface;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Log\LoggerInterface;
-use Psr\Log\LogLevel;
 
 /**
  * A class to log HTTP Requests and Responses of Guzzle.
+ *
+ * @author George Mponos <gmponos@gmail.com>
  */
 class LoggerMiddleware
 {
     /**
-     * @var bool Whether or not to log requests as they are made.
+     * Whether or not to log requests as they are made.
+     *
+     * @var bool
      */
     private $onExceptionOnly;
 
     /**
+     * Whether or not to log statistics.
+     *
      * @var bool
      */
     private $logStatistics;
 
     /**
-     * @var array
+     * @var HandlerInterface
      */
-    private $thresholds;
-
-    /**
-     * @var array
-     */
-    private $logCodeLevel = [];
-
-    /**
-     * @var bool
-     */
-    private $sensitive;
+    private $handler;
 
     /**
      * @var LoggerInterface
@@ -50,23 +46,20 @@ class LoggerMiddleware
      * Creates a callable middleware for logging requests and responses.
      *
      * @param LoggerInterface $logger
+     * @param HandlerInterface $handler
      * @param bool $onExceptionOnly The request and the response will be logged only in cases there is an exception or if they status code exceeds the thresholds.
      * @param bool $logStatistics If this is true an extra row will be added that will contain some HTTP statistics.
-     * @param array $thresholds
      */
     public function __construct(
         LoggerInterface $logger,
+        HandlerInterface $handler = null,
         $onExceptionOnly = false,
-        $logStatistics = false,
-        array $thresholds = []
+        $logStatistics = false
     ) {
         $this->logger = $logger;
         $this->onExceptionOnly = $onExceptionOnly;
         $this->logStatistics = $logStatistics;
-        $this->thresholds = array_merge([
-            'error' => 499,
-            'warning' => 399,
-        ], $thresholds);
+        $this->handler = $handler === null ? new ArrayHandler() : $handler;
     }
 
     /**
@@ -81,117 +74,36 @@ class LoggerMiddleware
             $this->setOptions($options);
 
             if ($this->onExceptionOnly === false) {
-                $this->logRequest($request);
+                $this->handler->log($this->logger, $request, $options);
                 if ($this->logStatistics && !isset($options['on_stats'])) {
-                    $options['on_stats'] = $this->logStatistics();
+                    $options['on_stats'] = function (TransferStats $stats) {
+                        $this->handler->log($this->logger, $stats);
+                    };
                 }
             }
 
-            return $handler($request, $options)
-                ->then(
-                    $this->handleSuccess($request),
-                    $this->handleFailure($request)
-                );
+            return $handler($request, $options)->then(
+                $this->handleSuccess($request, $options),
+                $this->handleFailure($request, $options)
+            );
         };
-    }
-
-    /**
-     * Returns the default log level for a response.
-     *
-     * @param RequestInterface|ResponseInterface|\Exception $message
-     * @return string LogLevel
-     */
-    private function getLogLevel($message = null)
-    {
-        if ($message === null || ($message instanceof \Exception)) {
-            return LogLevel::CRITICAL;
-        }
-
-        if ($message instanceof RequestInterface) {
-            return LogLevel::DEBUG;
-        }
-
-        if ($message instanceof ResponseInterface) {
-            $code = $message->getStatusCode();
-            if ($code === 0) {
-                return LogLevel::CRITICAL;
-            }
-
-            if (isset($this->logCodeLevel[$code])) {
-                return $this->logCodeLevel[$code];
-            }
-
-            if ($this->thresholds['error'] !== null && $code > $this->thresholds['error']) {
-                return LogLevel::CRITICAL;
-            }
-
-            if ($this->thresholds['warning'] !== null && $code > $this->thresholds['warning']) {
-                return LogLevel::ERROR;
-            }
-
-            return LogLevel::DEBUG;
-        }
-
-        throw new \InvalidArgumentException('Could not retrieve the log level because of unknown message class.');
-    }
-
-    /**
-     * @param RequestInterface $request
-     * @return void
-     */
-    private function logRequest(RequestInterface $request)
-    {
-        $this->logger->log(
-            $this->getLogLevel($request),
-            'Guzzle HTTP request',
-            $this->withRequestContext($request)
-        );
-    }
-
-    /**
-     * @return Closure
-     */
-    private function logStatistics()
-    {
-        return function (TransferStats $stats) {
-            $this->logger->debug('Guzzle HTTP statistics', [
-                'time' => $stats->getTransferTime(),
-                'uri' => $stats->getEffectiveUri(),
-            ]);
-        };
-    }
-
-    /**
-     * @param ResponseInterface $response
-     * @return void
-     */
-    private function logResponse(ResponseInterface $response)
-    {
-        $this->logger->log(
-            $this->getLogLevel($response),
-            'Guzzle HTTP response',
-            $this->withResponseContext($response)
-        );
     }
 
     /**
      * Returns a function which is handled when a request was successful.
      *
      * @param RequestInterface $request
+     * @param array $options
      * @return Closure
      */
-    private function handleSuccess(RequestInterface $request)
+    private function handleSuccess(RequestInterface $request, array $options)
     {
-        return function (ResponseInterface $response) use ($request) {
+        return function (ResponseInterface $response) use ($request, $options) {
+            // On exception only is true then it must not log the response since it was successful.
             if ($this->onExceptionOnly === false) {
-                $this->logResponse($response);
-                return $response;
+                $this->handler->log($this->logger, $response, $options);
             }
 
-            if ($response->getStatusCode() > $this->thresholds['warning']) {
-                $this->logRequest($request);
-                $this->logResponse($response);
-            }
             return $response;
         };
     }
@@ -200,108 +112,25 @@ class LoggerMiddleware
      * Returns a function which is handled when a request was rejected.
      *
      * @param RequestInterface $request
+     * @param array $options
      * @return Closure
      */
-    private function handleFailure(RequestInterface $request)
+    private function handleFailure(RequestInterface $request, array $options)
     {
-        return function (\Exception $reason) use ($request) {
+        return function (\Exception $reason) use ($request, $options) {
             if ($this->onExceptionOnly === true) {
-                $this->logRequest($request);
+                // This means that the request was not logged and since an exception happened we need to log the request too.
+                $this->handler->log($this->logger, $request, $options);
             }
 
             if ($reason instanceof RequestException && $reason->hasResponse()) {
-                $this->logResponse($reason->getResponse());
+                $this->handler->log($this->logger, $reason->getResponse(), $options);
                 return \GuzzleHttp\Promise\rejection_for($reason);
             }
 
-            $this->logger->log($this->getLogLevel($reason), 'Guzzle HTTP exception', $this->withReasonContext($reason));
+            $this->handler->log($this->logger, $reason, $options);
             return \GuzzleHttp\Promise\rejection_for($reason);
         };
-    }
-
-    /**
-     * Merges and return the response context
-     *
-     * @param \Exception $reason
-     * @param array $context
-     * @return array
-     */
-    private function withReasonContext(\Exception $reason, array $context = [])
-    {
-        $context['reason']['code'] = $reason->getCode();
-        $context['reason']['message'] = $reason->getMessage();
-        return $context;
-    }
-
-    /**
-     * Merges and return the request context
-     *
-     * @param RequestInterface $request
-     * @param array $context
-     * @return array
-     */
-    private function withRequestContext(RequestInterface $request, array $context = [])
-    {
-        $context['request']['method'] = $request->getMethod();
-        $context['request']['headers'] = $request->getHeaders();
-        $context['request']['uri'] = $request->getRequestTarget();
-        $context['request']['version'] = 'HTTP/' . $request->getProtocolVersion();
-
-        if ($request->getBody()->getSize() !== 0) {
-            $context['request']['body'] = $this->getBody($request);
-        }
-
-        return $context;
-    }
-
-    /**
-     * Merges and return the response context
-     *
-     * @param ResponseInterface $response
-     * @param array $context
-     * @return array
-     */
-    private function withResponseContext(ResponseInterface $response, array $context = [])
-    {
-        $context['response']['headers'] = $response->getHeaders();
-        $context['response']['statusCode'] = $response->getStatusCode();
-        $context['response']['version'] = 'HTTP/' . $response->getProtocolVersion();
-        $context['response']['message'] = $response->getReasonPhrase();
-
-        if ($response->getBody()->getSize() !== 0) {
-            $context['response']['body'] = $this->getBody($response);
-        }
-
-        return $context;
-    }
-
-    /**
-     * @param MessageInterface $message
-     * @return string
-     */
-    private function getBody(MessageInterface $message)
-    {
-        $stream = $message->getBody();
-        if ($stream->isSeekable() === false || $stream->isReadable() === false) {
-            return 'Body stream is not seekable/readable.';
-        }
-
-        if ($this->sensitive === true) {
-            return 'Body contains sensitive information therefore it is not included.';
-        }
-
-        if ($stream->getSize() >= 3500) {
-            return $stream->read(200) . ' (truncated...)';
-        }
-
-        $body = $stream->getContents();
-        $isJson = preg_grep('/application\/[\w\.\+]*(json)/', $message->getHeader('Content-Type'));
-        if (!empty($isJson)) {
-            $body = json_decode($body, true);
-        }
-
-        $stream->rewind();
-        return $body;
     }
 
     /**
@@ -315,28 +144,13 @@ class LoggerMiddleware
         }
 
         $options = $options['log'];
-        if (isset($options['requests'])) {
-            @trigger_error('Using option "requests" is deprecated and it will be removed on the next version. Use "on_exception_only"', E_USER_DEPRECATED);
-            if (!isset($options['on_exception_only'])) {
-                $options['on_exception_only'] = !$options['requests'];
-            }
-        }
 
-        $defaults = [
+        $options = array_merge([
             'on_exception_only' => $this->onExceptionOnly,
             'statistics' => $this->logStatistics,
-            'warning_threshold' => 399,
-            'error_threshold' => 499,
-            'levels' => [],
-            'sensitive' => false,
-        ];
+        ], $options);
 
-        $options = array_merge($defaults, $options);
-        $this->logCodeLevel = $options['levels'];
-        $this->thresholds['warning'] = $options['warning_threshold'];
-        $this->thresholds['error'] = $options['error_threshold'];
         $this->onExceptionOnly = $options['on_exception_only'];
         $this->logStatistics = $options['statistics'];
-        $this->sensitive = $options['sensitive'];
     }
 }