+++ CHANGELOG.md
@@ -2,11 +2,18 @@
 
 All Notable changes to `gmponos/guzzle_logger` will be documented in this file
 
+## 0.6.0 - 2018-10-15
+
+### Changed
+- **BREAKING CHANGE** Changed the constructor function of middleware.
+    - From now on you can pass a handler to it's constructor. Handlers are responsible for logging request/responses.
+    - Removed threshold argument.
+
 ## 0.5.0 - 2018-10-01
 
 ### Changed
 - **BREAKING CHANGE** Renamed the variable `$logRequestOnExceptionOnly` to `$onExceptionOnly`. The purpose of this constructor argument was 
-to log request and responses only if an exceptgition occurs. If you were manually setting this argument as true now you must set it
+to log request and responses only if an exception occurs. If you were manually setting this argument as true now you must set it
 as false as the variables meaning is inverted.
 - Deprecated the option `requests`. It will be removed on my next version.